requirements.txt
eventlet==0.32.0
python-socketio==5.4.1
